pyproject.toml
[project]
name = "llm-gradio-brochure-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.11.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
